Add new exercises solutions using vector concepts

+++ java/exercises-5/src/SomaVetor.java
@@ -7,17 +7,17 @@ public class SomaVetor {
     System.out.print("Quantos numeros voce vai digitar? ");
     int n = sc.nextInt();
     
-    int[] arr = new int[n];
+    double[] arr = new double[n];
 
     for(int i = 0; i < n; i++){
       System.out.print("Digite um nunero: ");
-      arr[i] = sc.nextInt();
+      arr[i] = sc.nextDouble();
     }
-    int total = arrSum(arr);
-    int med = total / arr.length;
+    double total = arrSum(arr);
+    double med = total / arr.length;
 
     System.out.print("VALORES = ");
-    for(int i : arr)
+    for(double i : arr)
       System.out.print(i + " ");
     
     System.out.println("\nSOMA = " + total);
@@ -26,10 +26,10 @@ public class SomaVetor {
     sc.close();
   }
 
-  public static int arrSum(int[] arr){
-    int soma = 0;
+  public static double arrSum(double[] arr){
+    double soma = 0;
 
-    for(int i : arr)
+    for(double i : arr)
       soma += i;
 
     return soma;
